Build log output, rename plugin => compiler
